Rename websocket.service to shell-websocket.service and add logger
This more clearly disambiguates this websocket service used for shells
from the websocket service used for ssm tunnels.
